Minor namespace fixup to be a wee bit more self-documenting of the intent.
Users of nupic should import validate from nupic.data.jsonhelpers, which may
also raise a ValidationError, also in nupic.data.jsonhelpers namespace, and
which is a subclass of ValueError.  With this changeset, it's obvious that
the ValidationError defined in nupic.data.jsonhelpers is explicitly the
implementation from Validictory.  Should we need to change it to a different
implementation, the assignment would change.